#0: move CB apis out of dataflow_api.h
